Removed unnecessary static cast to float within BotVisualizer::publishLoadRobot

+++ drake/systems/plants/BotVisualizer.h
@@ -122,7 +122,7 @@ class BotVisualizer {
             gdata.type = gdata.MESH;
             auto terrain = dynamic_cast<const DrakeShapes::HeightMapTerrain &>(geometry);
             gdata.num_float_data = 1;
-            gdata.float_data.push_back(static_cast<float>(1.0)); //scale
+            gdata.float_data.push_back(1.0f); //scale
             gdata.string_data = terrain.fname;
             break;
           }
